Revert "[Create] Skill UI"
This reverts commit b605ccfa7adcfc33429d75e9946f6b6b2cb6a5e0.
