--- Project_LoL/Assets/_Scripts/NHS/WeaponSystem/Skill.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 0347b60776eef974ba963ae0c53e0dd0
