--- Project_LoL/Assets/_Scripts/NHS/WeaponSystem/Skill_Data_SO.cs
@@ -1,10 +1,0 @@
-using UnityEngine;
-
-[CreateAssetMenu(fileName = "NewSkill", menuName = "Combat/Skill Data")]
-public class Skill_Data_SO : ScriptableObject
-{
-    public string skillName;
-    public Sprite icon;
-    public float  cooldown;
-    public int    maxAmmo;
-}
