--- Project_LoL/Assets/_Scripts/NHS/WeaponSystem/Skill_Data_SO.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 8239d75659760354ea2af62c0763a79d
