+++ Project_LoL/Assets/_Scripts/NHS/WeaponSystem/Weapon.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 643807fe767fa214cb6c94e383742a66
